% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_homer.R
\name{bed_to_homer}
\alias{bed_to_homer}
\title{Homer Ready Txt}
\usage{
bed_to_homer(bed, filename)
}
\arguments{
\item{bed}{bed formated file, first three column should be: chr, start, end , ...}

\item{filename}{either a character string naming a file or a connection open for writing}
}
\value{
None
}
\description{
This function takes a bed formatted file and creates HOMER annotation ready txt on MAC
}
\examples{
bed_to_homer("path/to/file", "path/to/save")

}
